requirements.txt
aiogram==2.23.1
revChatGPT==1.0.10
python-dotenv==0.21.0
